NODE-1088: Extract crypto.py
